requirements.txt
discord_webhook
python-dotenv